# AC.andNot(RC)