--- a/jmh/src/main/java/org/roaringbitmap/RoaringOnlyBenchmarkState.java
+++ b/jmh/src/main/java/org/roaringbitmap/RoaringOnlyBenchmarkState.java
@@ -38,7 +38,7 @@ public abstract class RoaringOnlyBenchmarkState {
     onlyArrayContainers = new ArrayList<>();
     onlyRunContainers = new ArrayList<>();
     onlyBitmapContainers = new ArrayList<>();
-    
+
     immutableBitmaps = new ArrayList<>();
     immutableOnlyArrayContainers = new ArrayList<>();
     immutableOnlyRunContainers = new ArrayList<>();