I've decided to try using `Util.advanceUntil` in `andNot` between `ArrayContainer` and `RunContainer`. The implementation present in the commit was verified with present tests and a new one, but for sure I'd test it more if going to be merged.

I run two benchmarks `part{3,4}_andNotArrayContainerVSRunContainerContainer` on `master` and the new branch.
```
BasicAndNotContainerBenchmark.part3_andNotArrayContainerVSRunContainerContainer  avgt   15  3327,444 ± 49,353  ns/op
BasicAndNotContainerBenchmark.part4_andNotArrayContainerVSRunContainerContainer  avgt   15  1877,370 ± 25,888  ns/op

BasicAndNotContainerBenchmark.part3_andNotArrayContainerVSRunContainerContainer  avgt   15  7790,214 ± 81,502  ns/op
BasicAndNotContainerBenchmark.part4_andNotArrayContainerVSRunContainerContainer  avgt   15  638,899 ± 4,385  ns/op
```

The code is slower for the present benchmark which `andNot` Array Container of 1024 elements with very dense Run Container of 1009 runs with 64512 cardinality.
The new benchmark uses Run Container with 5 very long runs and the new code is 3x faster.

I'm opening this Pull Request to discuss
- is it worth to use new approach?
- should there be a switch when to use old and new approach?

I think the new approach will be faster when there are long sequences of array elements covered by runs or between two runs.
If there are very few array elements per run or between runs the old approach should be faster.