@richardstartin 2017-06-13
Looks good to me but might it be worth using Thoughput and Sample bench mark modes since they paint a fuller picture? A Sample benchmark might have pointed to the speculative allocation of the short[] in the old code.