@lemire 2017-06-13
Any chance we might dynamically detect which code path is fastest?