@ppiotrow 2017-06-13
Richard, sure there will be more benchmarking. Not really understand "speculative allocation of the short[] as it occurs in both versions.

The bug is detected and I can fix it with next commit. As I marked the code is draft and requires to be properly tested.

Daniel, we may try to estimate how many array elements occur per a Run.
```java
int k = 10; //arbitraryEmelentsPerRun  - todo requires fine-tuning
if ((ac.cardinality/rc.numberOfRuns)>k) newCode() else oldCode()
```