@lemire 2017-06-13
Yes, so I suggest you look at such a heuristic to choose the fast code path quicky.