@richardstartin 2017-06-13
@ppiotrow 

On master I see unnecessary allocation for empty run containers:

```
233     short[] buffer = new short[cardinality]; 
234     if (x.nbrruns == 0) { 
235       return clone(); 
236     } 
```