@lemire 2017-07-07
@ppiotrow  Can you update us on the status of this PR?