@ppiotrow 2017-07-07
Sure, sorry for letting it hang without progress. I spent some time benchmarking and it actually showed that the improvement is not because of "smart search" by `advanceUntil` but by iterating over "runs" . I prepared method `iterateUntil` which - you know - continuously iterates ;) It appeared that it could almost replace `advanceUntil` even when the benchmark data was prepared to give boost for non-sequential search. I'll get back to it but recently I need to deal with non-development issues.