@lemire 2017-08-08
@ppiotrow Any update on this issue?