@ppiotrow 2017-08-29
@lemire Summer time slowly ends here and I've get back to that. I've noticed that there are no changes in performance which is caused by #178 .