@lemire 2017-08-29
Ok. Let us see if we can fix https://github.com/RoaringBitmap/RoaringBitmap/issues/178