@ppiotrow 2017-08-29
I've pushed rebased code to publish the recent changes. I've decided for now to switch to the `iterateUntil` as I can't clearly seem much improvement even in artificial benchmarks.

I'm looking for ideas how to benchmark the code. I've been using `RealDataBenchmarkAndNot.pairwiseAndNot` but it was not showing a difference even between hardcoded `byCardinalityAndNot` and `byRunsAndNot`.
The current code is fastest in both `part3` and `part4` version of `_andNotArrayContainerVSRunContainerContainer`, but the feature switch should be tuned/adjusted.