@lemire 2017-12-19
@ppiotrow Now that we appear to have solved https://github.com/RoaringBitmap/RoaringBitmap/issues/178 Can you have a look again?