@ppiotrow 2017-12-28
I'd like to get back to it again in January. I think that fixing #178 might be helpful, as I was not able to spot improvements in benchmarks where there should be some.