@ppiotrow 2018-01-27
I've force pushed the changes to this branch. They include the safe changes that could be merged now:
- removed speculative allocation of array containers
- andNot with "full" run container case
- removed need to decrement `i`
- various tests

I tried to benchmark the "by-runs" approach but still no luck (including the benchmark that extracts only array&run containers to `andNot` them).
If you accept the simple changes I'll port them to buffer implementation and we close it for now.