@ppiotrow 2018-01-27
Change of plans - I've just pushed the real data benchmarks and switched by default (no feature switch) to the `by-runs` implementation, which has better results. The results are for the newest code, I've just manually switched to one of by-runs or by-cardinality approach. I've run them on the official benchmark server via `./run.sh org.roaringbitmap.runcontainer.ArrayContainerAndNotRunContainerBenchmark` 

### by-cardinality only
```
Benchmark                                                                  (dataset)   Mode  Cnt      Score     Error  Units
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC           census-income  thrpt    6   4241.531 ±  15.334  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC              census1881  thrpt    6   3778.507 ±   7.461  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC           dimension_008  thrpt    6    825.106 ±   6.128  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC           dimension_003  thrpt    6    540.858 ±   2.518  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC           dimension_033  thrpt    6    793.983 ±   3.601  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC            uscensus2000  thrpt    6  26027.946 ± 192.707  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC         weather_sept_85  thrpt    6   1519.658 ±  13.538  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC      wikileaks-noquotes  thrpt    6   1974.499 ±  21.150  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC       census-income_srt  thrpt    6   1474.671 ±   1.775  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC          census1881_srt  thrpt    6  10918.254 ± 118.457  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC     weather_sept_85_srt  thrpt    6    975.964 ±   2.285  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC  wikileaks-noquotes_srt  thrpt    6   5456.867 ±  30.329  ops/s
```
### by-runs  only
```
Benchmark                                                                  (dataset)   Mode  Cnt      Score    Error  Units
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC           census-income  thrpt    6   4107.709 ±  9.203  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC              census1881  thrpt    6   4015.516 ± 57.414  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC           dimension_008  thrpt    6   1596.529 ± 39.238  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC           dimension_003  thrpt    6    546.158 ±  1.116  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC           dimension_033  thrpt    6   2211.556 ± 17.756  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC            uscensus2000  thrpt    6  26200.215 ± 29.353  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC         weather_sept_85  thrpt    6   1573.653 ± 15.746  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC      wikileaks-noquotes  thrpt    6   2921.767 ±  4.562  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC       census-income_srt  thrpt    6   2430.718 ±  6.003  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC          census1881_srt  thrpt    6  11145.742 ± 68.725  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC     weather_sept_85_srt  thrpt    6   1241.319 ±  4.702  ops/s
ArrayContainerAndNotRunContainerBenchmark.pairwiseACAndNotRC  wikileaks-noquotes_srt  thrpt    6   7610.497 ± 91.853  ops/s
```

In all benchmarks the by-runs approach has better throughput, especially `dimension_008`,`dimension_033` and `wikileaks-noquotes`. No improvement in ` uscensus2000` is because this is dead benchmark, no `andNot` is actually performed (could be explained by non overlapping containers etc).

I'm not using the `advanceUntil` binary search, the speed gain should be caused by better writes (`System.arrayCopy`). Instead of that I'm using `iterateUntil` that has different contract unfortunately
. I decided fit it *exactly* to my needs. Might be necessary to  remove it from very general class `Utils`.