@lemire 2018-01-28
The CI fails...

```
Tests in error: 
  TestUtil.testIterateUntil:96 » ArrayIndexOutOfBounds -1
```