@ppiotrow 2018-01-28
Right now it only fails on test coverage decreased. This is because there is a `byCardinality` method that is never used. It will be removed once we agree to always use `byRuns`.