@ppiotrow 2018-02-02
I've rebased once again as there were conflicts. The latest commits migrated the new `andNot` to the `Mappeable` containers.
Benchmark results:
### before
```
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC           census-income  thrpt    6   4115.511 ±  23.200  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC              census1881  thrpt    6   3700.978 ±  18.233  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC           dimension_008  thrpt    6    715.214 ±   2.405  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC           dimension_003  thrpt    6    392.583 ±   2.620  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC           dimension_033  thrpt    6    957.163 ±   1.972  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC            uscensus2000  thrpt    6  13781.241 ±  23.237  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC         weather_sept_85  thrpt    6   1542.546 ±  21.425  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC      wikileaks-noquotes  thrpt    6   1918.950 ±  13.967  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC       census-income_srt  thrpt    6   1633.116 ±   4.996  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC          census1881_srt  thrpt    6   7220.821 ± 118.992  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC     weather_sept_85_srt  thrpt    6    856.222 ±   5.260  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC  wikileaks-noquotes_srt  thrpt    6   4956.840 ±   6.646  ops/s
```
### after
```
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC           census-income  thrpt    6   3003.861 ± 11.617  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC              census1881  thrpt    6   3879.833 ± 23.042  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC           dimension_008  thrpt    6   1138.537 ± 12.382  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC           dimension_003  thrpt    6    424.930 ±  3.149  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC           dimension_033  thrpt    6   1964.210 ±  7.404  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC            uscensus2000  thrpt    6  13825.639 ± 61.090  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC         weather_sept_85  thrpt    6   1499.531 ± 10.418  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC      wikileaks-noquotes  thrpt    6   2041.218 ±  3.736  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC       census-income_srt  thrpt    6   1617.040 ±  5.009  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC          census1881_srt  thrpt    6   7130.179 ± 82.599  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC     weather_sept_85_srt  thrpt    6    876.837 ±  6.255  ops/s
ArrayContainerAndNotRunContainerBenchmark.immutablePairwiseACAndNotRC  wikileaks-noquotes_srt  thrpt    6   5662.352 ± 42.143  ops/s
```

I can see in results that some of them are faster while other slower. One more thing is that the benchmarks do not test off-heap. Have we ever tried the benchmarks for off heap?